norme(all): Standardize all files

+++ srcs/des/operation.c
@@ -1,12 +1,12 @@
 #include "../../incs/des.h"
 
-void permute(char *input, char *output, const u_int8_t *array, int size)
+void    permute(char *input, char *output, const u_int8_t *array, u_int8_t size)
 {
-	for (int count = 0; count < size; count++)
+	for (u_int8_t count = 0; count < size; count++)
 		output[count] = input[array[count] - 1];
 }
 
-void xor_bits_string(char *first, char *second, char *output, u_int64_t size)
+void    xor_bits_string(char *first, char *second, char *output, u_int64_t size)
 {
     for (u_int64_t count = 0; count < size; count++)
     {
